V 1.70 float изменён на decimal для лучшей точности

+++ backend/schemas/cart.py
@@ -1,4 +1,5 @@
 from typing import List
+from decimal import Decimal
 from pydantic import BaseModel, Field
 from schemas.product import ProductOut
 
@@ -6,13 +7,13 @@ from schemas.product import ProductOut
 class CartItemOut(BaseModel):
     product: ProductOut
     quantity: int
-    total_price: float
+    total_price: Decimal
 
 
 class CartOut(BaseModel):
     user_id: int
     cart_items: List[CartItemOut] = []
-    total_amount: float
+    total_amount: Decimal
 
 
 class CartItemCreate(BaseModel):
